Merge commit '34e4ddbfa39a176343bfb185458b7d58e06435ae' into develop

# Conflicts:
#	src/Atom_type/init.cpp
#	src/Makefile

+++ src/Atom_type/init.cpp
@@ -1,5 +1,4 @@
 #include "atom_type.h"
-#include <iostream>
 
 namespace sirius {
 
@@ -9,7 +8,7 @@ void Atom_type::init(int offset_lo__)
     if (initialized_) TERMINATE("can't initialize twice");
 
     offset_lo_ = offset_lo__;
-
+   
     /* read data from file if it exists */
     if (file_name_.length() > 0)
     {
@@ -40,7 +39,7 @@ void Atom_type::init(int offset_lo__)
 
             if (level.n != -1)
             {
-                for (int jst = 0; jst < (int)atomic_levels_.size(); jst++)
+                for (size_t jst = 0; jst < atomic_levels_.size(); jst++)
                 {
                     if ((atomic_levels_[jst].n == level.n) &&
                         (atomic_levels_[jst].l == level.l) &&
@@ -50,13 +49,13 @@ void Atom_type::init(int offset_lo__)
             }
         }
     }
-
+    
     /* check the nuclear charge */
     if (zn_ == 0) TERMINATE("zero atom charge");
 
     /* set default radial grid if it was not done by user */
     if (radial_grid_.num_points() == 0) set_radial_grid();
-
+    
     if (parameters_.esm_type() == full_potential_lapwlo)
     {
         /* initialize free atom density and potential */
@@ -65,7 +64,7 @@ void Atom_type::init(int offset_lo__)
         /* initialize aw descriptors if they were not set manually */
         if (aw_descriptors_.size() == 0) init_aw_descriptors(parameters_.lmax_apw());
 
-        if (static_cast<int>(aw_descriptors_.size()) != (parameters_.lmax_apw() + 1))
+        if (static_cast<int>(aw_descriptors_.size()) != (parameters_.lmax_apw() + 1)) 
             TERMINATE("wrong size of augmented wave descriptors");
 
         max_aw_order_ = 0;
@@ -84,18 +83,18 @@ void Atom_type::init(int offset_lo__)
             lo_descriptors_.push_back(lod);
         }
     }
-
+    
     /* initialize index of radial functions */
     indexr_.init(aw_descriptors_, lo_descriptors_);
 
     /* initialize index of muffin-tin basis functions */
     indexb_.init(indexr_);
-
+    
     /* get the number of core electrons */
     num_core_electrons_ = 0;
     if (parameters_.full_potential())
     {
-        for (int i = 0; i < (int)atomic_levels_.size(); i++)
+        for (size_t i = 0; i < atomic_levels_.size(); i++) 
         {
             if (atomic_levels_[i].core) num_core_electrons_ += atomic_levels_[i].occupancy;
         }
@@ -117,7 +116,7 @@ void Atom_type::init(int offset_lo__)
         for (int i2 = 0; i2 < indexr().size(); i2++)
         {
             int l2 = indexr(i2).l;
-
+            
             for (int i1 = 0; i1 <= i2; i1++)
             {
                 int l1 = indexr(i1).l;
@@ -133,9 +132,7 @@ void Atom_type::init(int offset_lo__)
             }
         }
     }
-
     idx_radial_integrals_ = mdarray<int, 2>(2, non_zero_elements.size());
-
     for (size_t j = 0; j < non_zero_elements.size(); j++)
     {
         idx_radial_integrals_(0, j) = non_zero_elements[j].first;
@@ -150,19 +147,15 @@ void Atom_type::init(int offset_lo__)
         rf_coef_ = mdarray<double, 3>(nullptr, num_mt_points_, 4, indexr().size());
         rf_coef_.allocate(1);
         rf_coef_.allocate_on_device();
-        vrf_coef_ = mdarray<double, 3>(nullptr, num_mt_points_, 4, lmmax_pot * indexr().size() * (parameters_.num_mag_dims() + 1));
+        vrf_coef_ = mdarray<double, 3>(nullptr, num_mt_points_, 4, lmmax_pot * indexr().size() * (parameters_.num_mag_dims() + 1)); 
         vrf_coef_.allocate(1);
         vrf_coef_.allocate_on_device();
         #else
         TERMINATE_NO_GPU
         #endif
     }
-
+    
     initialized_ = true;
 }
-
-
-
-
 
 }
